Move call_stack_historyt to its own file
I am about to add a few more kinds of history and having their own
files, like domains, probably makes more sense.
